Upgrade Docker image and workflow to Python 3.11
This commit updates the application's environment from Python 3.10 to 3.11 to use a more recent and supported version.

The `Dockerfile` is updated to use the `python-3.11` base image.

The GitHub Actions workflow for publishing to Docker Hub is simplified to only build and push the Python 3.11 image, which is now tagged as `latest`. The separate build step for Python 3.10 has been removed.